requirements.txt
django==5.0.7
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
tzlocal==5.2
gunicorn==22.0.0
psycopg2==2.9.9
flake8==7.1.0
pytest==8.3.2
